Issue #32: Re-architect filters to be plugins
