+++ server/scripts/tasks.py
@@ -1,4 +1,5 @@
 import requests
+
 from datetime import datetime, timedelta
 
 from twython import Twython
@@ -10,7 +11,7 @@ import csv
 from bs4 import BeautifulSoup
 import os
 from raven import Client
-
+import analyze_modules
 
 from ..models import User, FacebookAuth, TwitterAuth, Post
 from ..enums import GenderEnum
@@ -28,10 +29,15 @@ if config_map['prod'].SENTRY_DSN_WORKER:
     client = Client(config_map['prod'].SENTRY_DSN_WORKER)
 
 
-FACEBOOK_POSTS_FIELDS = ['id','caption','created_time','description','from{picture,name,gender}','icon','link','message','message_tags','name', 'object_id',
-                         'parent_id','permalink_url','picture','full_picture','place', 'properties', 'shares', 'source', 'status_type', 'story', 'story_tags' ,
-                         'type','updated_time','likes.summary(true)','reactions.summary(true)','comments.summary(true)']
-
+FACEBOOK_POSTS_FIELDS = ['id','caption','created_time','description',
+                         'from{picture,name,gender}','icon','link',
+                         'message','message_tags','name', 'object_id',
+                         'parent_id','permalink_url','picture','full_picture',
+                         'place', 'properties', 'shares', 'source',
+                         'status_type', 'story', 'story_tags' ,
+                         'type','updated_time','likes.summary(true)',
+                         'reactions.summary(true)','comments.summary(true)']
+ANALYSIS_TYPES = ['toxicity','gender_corporate','virality','news_score']
 FACEBOOK_URL = 'https://graph.facebook.com/v2.10/'
 
 basedir = os.path.abspath(os.path.dirname(__file__))
@@ -47,6 +53,7 @@ def get_posts_data_for_all_users(self):
             get_tweets_per_user.delay(user.id)
         if user.facebook_authorized:
             get_facebook_posts_per_user.delay(user.id)
+
 
 @celery.task(serializer='json', bind=True)
 def get_tweets_per_user(self, user_id):
@@ -190,119 +197,12 @@ def _add_news_post(post, source, quintile):
 
 @celery.task(serializer='json', bind=True)
 def analyze_post(self, post_id):
-    analyze_toxicity(post_id)
-    analyze_gender_corporate(post_id)
-    analyze_virality(post_id)
-    get_news_score(post_id)
-
-
-def analyze_toxicity(post_id):
-    post = Post.query.get(post_id)
-    if not post or post.has_toxicity_rate():
-        logger.warning("post {} doesn't exist or already has toxicity rate".format(post_id))
-        return
-    text = post.get_text()
-
-    # Generates API client object dynamically based on service name and version.
-    # cache_dicovery=False to silence google file_cache error https://github.com/google/google-api-python-client/issues/299
-    service = discovery.build('commentanalyzer', 'v1alpha1', developerKey=current_app.config['GOOGLE_API_KEY'], cache_discovery=False)
-
-    analyze_request = {
-        'comment': {'text': text},
-        'requestedAttributes': {'TOXICITY': {}},
-        'doNotStore': True
-    }
-
-    try:
-        response = service.comments().analyze(body=analyze_request).execute()
-        score = response["attributeScores"]["TOXICITY"]["summaryScore"]["value"]
-    except:
-        logger.info('could not get toxicity score for post {}'.format(post_id))
-        score = -1
-    post.update_toxicity(score)
-
-def analyze_gender_corporate(post_id):
-    post = Post.query.get(post_id)
-    if not post or post.has_gender_corporate():
-        logger.warning("post {} doesn't exist or already has gender/corporate".format(post_id))
-        return
-    is_facebook = post.source=='facebook'
-    gender = GenderEnum.unknown
-    corporate = False
-    if is_facebook and 'gender' in post.content['from']:
-        gender = GenderEnum.fromString(post.content['from']['gender'])
-    else:
-        result, conf = name_classifier.predictGenderbyName(post.get_author_name())
-        #score = name_gender_analyzer.process(post.get_author_name())
-        gender = GenderEnum.fromString(result)
-    if post.is_news:
-        gender = GenderEnum.unknown
-    if gender==GenderEnum.unknown or (is_facebook and 'category' in post.content['from']):
-        corporate = True
-    post.update_gender_corporate(gender, corporate)
-
-def analyze_virality(post_id):
-    post = Post.query.get(post_id)
-    if not post or post.has_virality():
-        logger.warning("post {} doesn't exist or already has virality".format(post_id))
-        return
-    is_facebook = post.source=='facebook'
-
-    likes = post.content['reactions']['summary']['total_count'] if is_facebook else post.content['favorite_count']
-    if is_facebook:
-        comments = post.content['comments']['summary']['total_count']
-    else:
-        comments = count_tweet_replies(post.content)
-        post.update_replies_count(comments)
-    shares = 0
-    if is_facebook:
-        if 'shares' in post.content:
-            shares = post.content['shares']['count']
-    else:
-        shares = post.content['retweet_count']
-
-    total_reaction = likes+shares+comments
-    post.virality_count = max(post.virality_count, total_reaction)
-    db.session.commit()
-
-def get_news_score(post_id):
-    post = Post.query.get(post_id)
-    if not post or post.has_news_score():
-        logger.warning("post {} doesn't exist or already has news score".format(post_id))
-        return
-
-    is_facebook = post.source=='facebook'
-    score = 0
-    if post.has_link:
-        urls = [post.content['link']] if is_facebook else [x['expanded_url'] for x in post.content['entities']['urls']]
-        for url in urls:
-            try:
-                html = urllib.urlopen(url).read()
-            except:
-                html = ""
-            soup = BeautifulSoup(html, "html.parser")
-            # kill all script and style elements
-            for script in soup(["script", "style"]):
-                script.extract()  # rip it out
-            # get text
-            text = soup.get_text()
-            r = requests.post(current_app.config['NEWS_LABELLER_URL']+'/predict.json', json = {'text':text})
-            result = r.json()
-            if 'taxonomies' in result:
-                scores = [float(x['score']) for x in result['taxonomies'] if '/news' in x['label'].lower()]
-                scores.append(score)
-                score = max(scores)
-    else:
-        text = post.get_text()
-        r = requests.post(current_app.config['NEWS_LABELLER_URL']+'/predict.json', json = {'text':text})
-        result = r.json()
-        if 'taxonomies' in result:
-            scores = [float(x['score'])for x in result['taxonomies'] if '/news' in x['label'].lower()]
-            score = max(scores) if len(scores)>0 else 0
-    if post.is_news:
-        score = min(1, score+0.6)
-    post.news_score = score
-    db.session.commit()
+    for analysis_type in ANALYSIS_TYPES:
+        getattr(analyze_modules,"analyze_%s"%(analysis_type))()
+    #analyze_toxicity(post_id)
+    #analyze_gender_corporate(post_id)
+    #analyze_virality(post_id)
+    #analyze_news_score(post_id)
